+++ frontend/src/app/sdk/sdk.component.ts
@@ -1,10 +1,9 @@
 import { Component, OnInit } from '@angular/core';
-import { LoginService } from '../login.service';
 
 @Component({
-  selector: 'app-sdk',
-  templateUrl: './sdk.component.html',
-  styleUrls: ['./sdk.component.scss']
+    selector: 'app-sdk',
+    templateUrl: './sdk.component.html',
+    styleUrls: ['./sdk.component.scss']
 })
 export class SdkComponent implements OnInit {
 
@@ -30,9 +29,9 @@ export class SdkComponent implements OnInit {
         }
     ];
 
-  constructor(public loginService: LoginService) { }
+    constructor() { }
 
-  ngOnInit() {
-  }
+    ngOnInit() {
+    }
 
 }